+++ src/Screens/Flows/App/PrivacyScreen/Hooks/useBackupDetails.tsx
@@ -5,14 +5,18 @@ import { LocalDevice } from "~Model"
 type Props = {
     closePasswordPrompt: () => void
     openWalletMgmtSheetWithDelay: (delay: number) => void
-    openBackupPhraseSheetWithDelay: (delay: number, mnemonicArray: string[], deviceToBackup: LocalDevice) => void
+    openBackupPhraseSheetWithDelay: (
+        delay: number,
+        backupDetails: string[] | string,
+        deviceToBackup: LocalDevice,
+    ) => void
     openPasswordPrompt: () => void
     closeWalletMgmtSheet: () => void
     devices: LocalDevice[]
     isWalletSecurityBiometrics: boolean
 }
 
-export const useBackupMnemonic = ({
+export const useBackupDetails = ({
     closePasswordPrompt,
     openWalletMgmtSheetWithDelay,
     openBackupPhraseSheetWithDelay,
@@ -42,6 +46,9 @@ export const useBackupMnemonic = ({
                 if (wallet?.mnemonic) {
                     setDeviceToBackup(devices[0])
                     openBackupPhraseSheetWithDelay(300, wallet.mnemonic, devices[0])
+                } else if (wallet?.privateKey) {
+                    setDeviceToBackup(devices[0])
+                    openBackupPhraseSheetWithDelay(300, wallet.privateKey, devices[0])
                 }
             }
         } else {
@@ -73,10 +80,12 @@ export const useBackupMnemonic = ({
 
                 if (wallet?.mnemonic?.length) {
                     openBackupPhraseSheetWithDelay(300, wallet.mnemonic, deviceToBackup)
+                } else if (wallet?.privateKey) {
+                    openBackupPhraseSheetWithDelay(300, wallet.privateKey, deviceToBackup)
                 }
             }
         },
-        [closePasswordPrompt, openBackupPhraseSheetWithDelay, deviceToBackup],
+        [closePasswordPrompt, deviceToBackup, openBackupPhraseSheetWithDelay],
     )
 
     /*
@@ -95,6 +104,8 @@ export const useBackupMnemonic = ({
 
                 if (wallet?.mnemonic) {
                     openBackupPhraseSheetWithDelay(300, wallet.mnemonic, device)
+                } else if (wallet?.privateKey) {
+                    openBackupPhraseSheetWithDelay(300, wallet.privateKey, device)
                 }
             } else {
                 openPasswordPrompt()