+++ src/Screens/Flows/App/SwapScreen/SwapScreen.tsx
@@ -13,7 +13,7 @@ import {
     SelectAccountBottomSheet,
 } from "~Components"
 import { AnalyticsEvent, DiscoveryDApp } from "~Constants"
-import { useAnalyticTracking, useBottomSheetModal, useThemedStyles } from "~Hooks"
+import { useAnalyticTracking, useBottomSheetModal, useSetSelectedAccount, useThemedStyles } from "~Hooks"
 import { useFetchFeaturedDApps } from "~Hooks/useFetchFeaturedDApps"
 import { useI18nContext } from "~i18n"
 import { RootStackParamListHome, Routes } from "~Navigation"
@@ -24,7 +24,6 @@ import {
     selectSelectedAccount,
     selectSwapFeaturedDapps,
     selectVisibleAccounts,
-    setSelectedAccount,
     useAppDispatch,
     useAppSelector,
 } from "~Storage/Redux"
@@ -44,6 +43,8 @@ export const SwapScreen = () => {
     const accounts = useAppSelector(selectVisibleAccounts)
     const swappDApps = useAppSelector(selectSwapFeaturedDapps)
     const bookmarkedDApps = useAppSelector(selectBookmarkedDapps)
+
+    const { onSetSelectedAccount } = useSetSelectedAccount()
 
     const { isFetching } = useFetchFeaturedDApps()
 
@@ -127,7 +128,7 @@ export const SwapScreen = () => {
                     <SelectAccountBottomSheet
                         closeBottomSheet={closeSelectAccountBottonSheet}
                         accounts={accounts}
-                        setSelectedAccount={setSelectedAccount}
+                        setSelectedAccount={onSetSelectedAccount}
                         selectedAccount={selectedAccount}
                         isBalanceVisible={isBalanceVisible}
                         ref={selectAccountBottomSheetRef}