+++ src/Storage/Redux/Migrations/index.ts
@@ -6,6 +6,8 @@ import { MigrationManifest } from "redux-persist/es/types"
 import { Migration6 } from "./Migration6"
 import { Migration7 } from "./Migration7"
 import { Migration8 } from "./Migration8"
+import { Migration9 } from "./Migration9"
+import { Migration10 } from "./Migration10"
 
 export const migrationUpdates: MigrationManifest = {
     2: state => Migration2(state),
@@ -15,4 +17,6 @@ export const migrationUpdates: MigrationManifest = {
     6: state => Migration6(state),
     7: state => Migration7(state),
     8: state => Migration8(state),
+    9: state => Migration9(state),
+    10: state => Migration10(state),
 }