+++ src/Utils/ReanimatedUtils/Reanimated.ts
@@ -21,7 +21,8 @@ export function numberToPercentWorklet(
     options: {
         precision: number
         absolute: boolean
-    } = { precision: DEFAULT_PRECISION, absolute: DEFAULT_ABSOLUTE },
+        locale?: Intl.LocalesArgument
+    } = { precision: DEFAULT_PRECISION, absolute: DEFAULT_ABSOLUTE, locale: "en-US" },
 ): string {
     "worklet"
 
@@ -30,11 +31,17 @@ export function numberToPercentWorklet(
         return Math.round(v * p) / p
     }
 
-    const { precision, absolute } = options
+    const { precision, absolute, locale } = options
 
     if (precision < 0) {
         throw new Error("numberToPercentWorklet does not handle negative precision values")
     }
+
+    const formatter = new Intl.NumberFormat(locale?.toString(), {
+        style: "decimal",
+        useGrouping: true,
+        minimumFractionDigits: precision,
+    })
 
     if (value === undefined || isNaN(value)) {
         return "-"
@@ -48,21 +55,8 @@ export function numberToPercentWorklet(
 
     // return raw value when precision is zero
     if (precision === 0) {
-        return `${shapedValue}%`
+        return `${formatter.format(shapedValue)}%`
     }
 
-    // Add trailing zeros when value has no decimal
-    if (Number.isInteger(shapedValue)) {
-        return `${shapedValue}.${"0".repeat(precision)}%`
-    }
-
-    let endingZeros = ""
-    const shiftedValue = shapedValue * Math.pow(10, precision)
-    for (let d = 0; d < precision; d++) {
-        if (shiftedValue % Math.pow(10, d + 1) === 0) {
-            endingZeros += "0"
-        }
-    }
-
-    return `${shapedValue}${endingZeros}%`
+    return `${formatter.format(shapedValue)}%`
 }