+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 
 
+## 2.0.8 (2025-02-13)
+
+
+### Bug Fixes
+
+* fontsize and weight on token cards and assets detail page ([#2685](https://github.com/vechain/veworld-mobile/issues/2685)) ([20373da](https://github.com/vechain/veworld-mobile/commit/20373da32fb116c8fe1f79a1f6554b1becdc395e))
+
+## 2.0.8 (2025-02-04)
+
+
+### Bug Fixes
+
+* updated copy and removed fiat balance in wallet details accounts ([#2641](https://github.com/vechain/veworld-mobile/issues/2641)) ([ac051b3](https://github.com/vechain/veworld-mobile/commit/ac051b37eae52c73a05b7d2501fe1ce43960c0d3))
+
 ## 2.0.7 (2025-01-24)